Move message processing to a helper function
Move processing of messages to a helper function so the code isn't as
complex in the main program.
