+++ lib/helpers.js
@@ -4,7 +4,7 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.generateDebugString = exports.getDebugInfo = undefined;
+exports.processMessages = exports.generateDebugString = exports.getDebugInfo = undefined;
 
 var getDebugInfo = exports.getDebugInfo = function () {
   var _ref = _asyncToGenerator(function* (worker) {
@@ -56,10 +56,120 @@ var generateDebugString = exports.generateDebugString = function () {
   };
 }();
 
+var processMessages = exports.processMessages = function () {
+  var _ref3 = _asyncToGenerator(function* (response, textEditor, showRule) {
+    return Promise.all(response.map(function () {
+      var _ref4 = _asyncToGenerator(function* (_ref5) {
+        var message = _ref5.message,
+            line = _ref5.line,
+            severity = _ref5.severity,
+            ruleId = _ref5.ruleId,
+            column = _ref5.column,
+            fix = _ref5.fix,
+            endLine = _ref5.endLine,
+            endColumn = _ref5.endColumn;
+
+        var filePath = textEditor.getPath();
+        var textBuffer = textEditor.getBuffer();
+        var linterFix = null;
+        if (fix) {
+          var fixRange = new _atom.Range(textBuffer.positionForCharacterIndex(fix.range[0]), textBuffer.positionForCharacterIndex(fix.range[1]));
+          linterFix = {
+            range: fixRange,
+            newText: fix.text
+          };
+        }
+        var msgLine = line - 1;
+        var msgCol = void 0;
+        var msgEndLine = void 0;
+        var msgEndCol = void 0;
+        var eslintFullRange = false;
+        if (typeof endColumn !== 'undefined' && typeof endLine !== 'undefined') {
+          eslintFullRange = true;
+          // Here we always want the column to be a number
+          msgCol = Math.max(0, column - 1);
+          msgEndLine = endLine - 1;
+          msgEndCol = endColumn - 1;
+        } else {
+          // We want msgCol to remain undefined if it was initially so
+          // `rangeFromLineNumber` will give us a range over the entire line
+          msgCol = typeof column !== 'undefined' ? column - 1 : column;
+        }
+
+        var ret = void 0;
+        var range = void 0;
+        try {
+          if (eslintFullRange) {
+            validatePoint(textEditor, msgLine, msgCol);
+            validatePoint(textEditor, msgEndLine, msgEndCol);
+            range = [[msgLine, msgCol], [msgEndLine, msgEndCol]];
+          } else {
+            range = (0, _atomLinter.rangeFromLineNumber)(textEditor, msgLine, msgCol);
+          }
+          ret = {
+            filePath: filePath,
+            type: severity === 1 ? 'Warning' : 'Error',
+            range: range
+          };
+
+          if (showRule) {
+            var elName = ruleId ? 'a' : 'span';
+            var href = ruleId ? ' href=' + (0, _eslintRuleDocumentation2.default)(ruleId).url : '';
+            ret.html = '<' + elName + href + ' class="badge badge-flexible eslint">' + ((ruleId || 'Fatal') + '</' + elName + '> ' + (0, _escapeHtml2.default)(message));
+          } else {
+            ret.text = message;
+          }
+          if (linterFix) {
+            ret.fix = linterFix;
+          }
+        } catch (err) {
+          var errMsgRange = msgLine + 1 + ':' + msgCol;
+          if (eslintFullRange) {
+            errMsgRange += ' - ' + (msgEndLine + 1) + ':' + (msgEndCol + 1);
+          }
+          var rangeText = 'Requested ' + (eslintFullRange ? 'start point' : 'range') + ': ' + errMsgRange;
+          var issueURL = 'https://github.com/AtomLinter/linter-eslint/issues/new';
+          var titleText = 'Invalid position given by \'' + ruleId + '\'';
+          var title = encodeURIComponent(titleText);
+          var body = encodeURIComponent(['ESLint returned a point that did not exist in the document being edited.', 'Rule: ' + ruleId, rangeText, '', '', '<!-- If at all possible, please include code to reproduce this issue! -->', '', '', 'Debug information:', '```json', JSON.stringify((yield getDebugInfo()), null, 2), '```'].join('\n'));
+          var newIssueURL = issueURL + '?title=' + title + '&body=' + body;
+          ret = {
+            type: 'Error',
+            severity: 'error',
+            html: (0, _escapeHtml2.default)(titleText) + '. See the trace for details. ' + ('<a href="' + newIssueURL + '">Report this!</a>'),
+            filePath: filePath,
+            range: (0, _atomLinter.rangeFromLineNumber)(textEditor, 0),
+            trace: [{
+              type: 'Trace',
+              text: 'Original message: ' + ruleId + ' - ' + message,
+              filePath: filePath,
+              severity: 'info'
+            }, {
+              type: 'Trace',
+              text: rangeText,
+              filePath: filePath,
+              severity: 'info'
+            }]
+          };
+        }
+
+        return ret;
+      });
+
+      return function (_x8) {
+        return _ref4.apply(this, arguments);
+      };
+    }()));
+  });
+
+  return function processMessages(_x5, _x6, _x7) {
+    return _ref3.apply(this, arguments);
+  };
+}();
+
 exports.spawnWorker = spawnWorker;
 exports.showError = showError;
 exports.idsToIgnoredRules = idsToIgnoredRules;
-exports.validatePoint = validatePoint;
 
 var _child_process = require('child_process');
 
@@ -68,6 +178,16 @@ var _child_process2 = _interopRequireDefault(_child_process);
 var _processCommunication = require('process-communication');
 
 var _path = require('path');
+
+var _escapeHtml = require('escape-html');
+
+var _escapeHtml2 = _interopRequireDefault(_escapeHtml);
+
+var _eslintRuleDocumentation = require('eslint-rule-documentation');
+
+var _eslintRuleDocumentation2 = _interopRequireDefault(_eslintRuleDocumentation);
+
+var _atomLinter = require('atom-linter');
 
 var _atom = require('atom');
