+++ lib/main.js
@@ -1,24 +1,13 @@
 'use strict';
 'use babel';
 
-var _escapeHtml = require('escape-html');
-
-var _escapeHtml2 = _interopRequireDefault(_escapeHtml);
-
-var _eslintRuleDocumentation = require('eslint-rule-documentation');
-
-var _eslintRuleDocumentation2 = _interopRequireDefault(_eslintRuleDocumentation);
+// eslint-disable-next-line import/no-extraneous-dependencies, import/extensions
 
 var _atom = require('atom');
 
 var _helpers = require('./helpers');
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
 function _asyncToGenerator(fn) { return function () { var gen = fn.apply(this, arguments); return new Promise(function (resolve, reject) { function step(key, arg) { try { var info = gen[key](arg); var value = info.value; } catch (error) { reject(error); return; } if (info.done) { resolve(value); } else { return Promise.resolve(value).then(function (value) { step("next", value); }, function (err) { step("throw", err); }); } } return step("next"); }); }; }
-
-// eslint-disable-next-line import/no-extraneous-dependencies, import/extensions
-
 
 // Configuration
 var scopes = [];
@@ -117,10 +106,9 @@ module.exports = {
     }));
 
     var initializeWorker = function initializeWorker() {
-      var _spawnWorker = (0, _helpers.spawnWorker)();
-
-      var worker = _spawnWorker.worker;
-      var subscription = _spawnWorker.subscription;
+      var _spawnWorker = (0, _helpers.spawnWorker)(),
+          worker = _spawnWorker.worker,
+          subscription = _spawnWorker.subscription;
 
       _this.worker = worker;
       _this.subscriptions.add(subscription);
@@ -139,8 +127,6 @@ module.exports = {
   },
   provideLinter: function provideLinter() {
     var _this2 = this;
-
-    var Helpers = require('atom-linter');
 
     return {
       name: 'ESLint',
@@ -176,111 +162,7 @@ module.exports = {
              */
             return null;
           }
-          var messagePromises = response.map(function () {
-            var _ref2 = _asyncToGenerator(function* (_ref3) {
-              var message = _ref3.message;
-              var line = _ref3.line;
-              var severity = _ref3.severity;
-              var ruleId = _ref3.ruleId;
-              var column = _ref3.column;
-              var fix = _ref3.fix;
-              var endLine = _ref3.endLine;
-              var endColumn = _ref3.endColumn;
-
-              var textBuffer = textEditor.getBuffer();
-              var linterFix = null;
-              if (fix) {
-                var fixRange = new _atom.Range(textBuffer.positionForCharacterIndex(fix.range[0]), textBuffer.positionForCharacterIndex(fix.range[1]));
-                linterFix = {
-                  range: fixRange,
-                  newText: fix.text
-                };
-              }
-              var msgLine = line - 1;
-              var msgCol = void 0;
-              var msgEndLine = void 0;
-              var msgEndCol = void 0;
-              var eslintFullRange = false;
-              if (typeof endColumn !== 'undefined' && typeof endLine !== 'undefined') {
-                eslintFullRange = true;
-                // Here we always want the column to be a number
-                msgCol = Math.max(0, column - 1);
-                msgEndLine = endLine - 1;
-                msgEndCol = endColumn - 1;
-              } else {
-                // We want msgCol to remain undefined if it was initially so
-                // `rangeFromLineNumber` will give us a range over the entire line
-                msgCol = typeof column !== 'undefined' ? column - 1 : column;
-              }
-
-              var ret = void 0;
-              var range = void 0;
-              try {
-                if (eslintFullRange) {
-                  (0, _helpers.validatePoint)(textEditor, msgLine, msgCol);
-                  (0, _helpers.validatePoint)(textEditor, msgEndLine, msgEndCol);
-                  range = [[msgLine, msgCol], [msgEndLine, msgEndCol]];
-                } else {
-                  range = Helpers.rangeFromLineNumber(textEditor, msgLine, msgCol);
-                }
-                ret = {
-                  filePath: filePath,
-                  type: severity === 1 ? 'Warning' : 'Error',
-                  range: range
-                };
-
-                if (showRule) {
-                  var elName = ruleId ? 'a' : 'span';
-                  var href = ruleId ? ' href=' + (0, _eslintRuleDocumentation2.default)(ruleId).url : '';
-                  ret.html = '<' + elName + href + ' class="badge badge-flexible eslint">' + ((ruleId || 'Fatal') + '</' + elName + '> ' + (0, _escapeHtml2.default)(message));
-                } else {
-                  ret.text = message;
-                }
-                if (linterFix) {
-                  ret.fix = linterFix;
-                }
-              } catch (err) {
-                var errMsgRange = msgLine + 1 + ':' + msgCol;
-                if (eslintFullRange) {
-                  errMsgRange += ' - ' + (msgEndLine + 1) + ':' + (msgEndCol + 1);
-                }
-                var rangeText = 'Requested ' + (eslintFullRange ? 'start point' : 'range') + ': ' + errMsgRange;
-                var issueURL = 'https://github.com/AtomLinter/linter-eslint/issues/new';
-                var titleText = 'Invalid position given by \'' + ruleId + '\'';
-                var title = encodeURIComponent(titleText);
-                var body = encodeURIComponent(['ESLint returned a point that did not exist in the document being edited.', 'Rule: ' + ruleId, rangeText, '', '', '<!-- If at all possible, please include code to reproduce this issue! -->', '', '', 'Debug information:', '```json', JSON.stringify((yield (0, _helpers.getDebugInfo)()), null, 2), '```'].join('\n'));
-                var newIssueURL = issueURL + '?title=' + title + '&body=' + body;
-                ret = {
-                  type: 'Error',
-                  severity: 'error',
-                  html: (0, _escapeHtml2.default)(titleText) + '. See the trace for details. ' + ('<a href="' + newIssueURL + '">Report this!</a>'),
-                  filePath: filePath,
-                  range: Helpers.rangeFromLineNumber(textEditor, 0),
-                  trace: [{
-                    type: 'Trace',
-                    text: 'Original message: ' + ruleId + ' - ' + message,
-                    filePath: filePath,
-                    severity: 'info'
-                  }, {
-                    type: 'Trace',
-                    text: rangeText,
-                    filePath: filePath,
-                    severity: 'info'
-                  }]
-                };
-              }
-
-              return ret;
-            });
-
-            return function (_x) {
-              return _ref2.apply(this, arguments);
-            };
-          }());
-
-          return Promise.all(messagePromises).then(function (messages) {
-            return messages;
-          });
+          return (0, _helpers.processMessages)(response, textEditor, showRule);
         });
       }
     };
