+++ lib/worker-helpers.js
@@ -97,9 +97,8 @@ function findESLintDirectory(modulesDir, config, projectPath) {
 }
 
 function getESLintFromDirectory(modulesDir, config, projectPath) {
-  var _findESLintDirectory = findESLintDirectory(modulesDir, config, projectPath);
-
-  var ESLintDirectory = _findESLintDirectory.path;
+  var _findESLintDirectory = findESLintDirectory(modulesDir, config, projectPath),
+      ESLintDirectory = _findESLintDirectory.path;
 
   try {
     // eslint-disable-next-line import/no-dynamic-require
