+++ lib/worker.js
@@ -51,12 +51,12 @@ function fixJob(argv, eslint) {
 }
 
 (0, _processCommunication.create)().onRequest('job', function (_ref, job) {
-  var contents = _ref.contents;
-  var type = _ref.type;
-  var config = _ref.config;
-  var filePath = _ref.filePath;
-  var projectPath = _ref.projectPath;
-  var rules = _ref.rules;
+  var contents = _ref.contents,
+      type = _ref.type,
+      config = _ref.config,
+      filePath = _ref.filePath,
+      projectPath = _ref.projectPath,
+      rules = _ref.rules;
 
   global.__LINTER_ESLINT_RESPONSE = [];
