+++ src/helpers.js
@@ -3,9 +3,12 @@
 import ChildProcess from 'child_process'
 import { createFromProcess } from 'process-communication'
 import { join } from 'path'
+import escapeHTML from 'escape-html'
+import ruleURI from 'eslint-rule-documentation'
+import { rangeFromLineNumber } from 'atom-linter'
 
 // eslint-disable-next-line import/no-extraneous-dependencies, import/extensions
-import { Disposable } from 'atom'
+import { Disposable, Range } from 'atom'
 
 const RULE_OFF_SEVERITY = 0
 
@@ -57,7 +60,7 @@ export function idsToIgnoredRules(ruleIds = []) {
   }, {})
 }
 
-export function validatePoint(textEditor, line, col) {
+function validatePoint(textEditor, line, col) {
   const buffer = textEditor.getBuffer()
   // Clip the given point to a valid one, and check if it equals the original
   if (!buffer.clipPosition([line, col]).isEqual([line, col])) {
@@ -109,4 +112,115 @@ export async function generateDebugString(worker) {
     `linter-eslint configuration: ${JSON.stringify(debug.linterEslintConfig, null, 2)}`
   ]
   return details.join('\n')
+}
+
+export async function processMessages(response, textEditor, showRule) {
+  return Promise.all(response.map(async ({
+    message, line, severity, ruleId, column, fix, endLine, endColumn
+  }) => {
+    const filePath = textEditor.getPath()
+    const textBuffer = textEditor.getBuffer()
+    let linterFix = null
+    if (fix) {
+      const fixRange = new Range(
+        textBuffer.positionForCharacterIndex(fix.range[0]),
+        textBuffer.positionForCharacterIndex(fix.range[1])
+      )
+      linterFix = {
+        range: fixRange,
+        newText: fix.text
+      }
+    }
+    const msgLine = line - 1
+    let msgCol
+    let msgEndLine
+    let msgEndCol
+    let eslintFullRange = false
+    if (typeof endColumn !== 'undefined' && typeof endLine !== 'undefined') {
+      eslintFullRange = true
+      // Here we always want the column to be a number
+      msgCol = Math.max(0, column - 1)
+      msgEndLine = endLine - 1
+      msgEndCol = endColumn - 1
+    } else {
+      // We want msgCol to remain undefined if it was initially so
+      // `rangeFromLineNumber` will give us a range over the entire line
+      msgCol = typeof column !== 'undefined' ? column - 1 : column
+    }
+
+    let ret
+    let range
+    try {
+      if (eslintFullRange) {
+        validatePoint(textEditor, msgLine, msgCol)
+        validatePoint(textEditor, msgEndLine, msgEndCol)
+        range = [[msgLine, msgCol], [msgEndLine, msgEndCol]]
+      } else {
+        range = rangeFromLineNumber(textEditor, msgLine, msgCol)
+      }
+      ret = {
+        filePath,
+        type: severity === 1 ? 'Warning' : 'Error',
+        range
+      }
+
+      if (showRule) {
+        const elName = ruleId ? 'a' : 'span'
+        const href = ruleId ? ` href=${ruleURI(ruleId).url}` : ''
+        ret.html = `<${elName}${href} class="badge badge-flexible eslint">` +
+          `${ruleId || 'Fatal'}</${elName}> ${escapeHTML(message)}`
+      } else {
+        ret.text = message
+      }
+      if (linterFix) {
+        ret.fix = linterFix
+      }
+    } catch (err) {
+      let errMsgRange = `${msgLine + 1}:${msgCol}`
+      if (eslintFullRange) {
+        errMsgRange += ` - ${msgEndLine + 1}:${msgEndCol + 1}`
+      }
+      const rangeText = `Requested ${eslintFullRange ? 'start point' : 'range'}: ${errMsgRange}`
+      const issueURL = 'https://github.com/AtomLinter/linter-eslint/issues/new'
+      const titleText = `Invalid position given by '${ruleId}'`
+      const title = encodeURIComponent(titleText)
+      const body = encodeURIComponent([
+        'ESLint returned a point that did not exist in the document being edited.',
+        `Rule: ${ruleId}`,
+        rangeText,
+        '', '',
+        '<!-- If at all possible, please include code to reproduce this issue! -->',
+        '', '',
+        'Debug information:',
+        '```json',
+        JSON.stringify(await getDebugInfo(), null, 2),
+        '```'
+      ].join('\n'))
+      const newIssueURL = `${issueURL}?title=${title}&body=${body}`
+      ret = {
+        type: 'Error',
+        severity: 'error',
+        html: `${escapeHTML(titleText)}. See the trace for details. ` +
+          `<a href="${newIssueURL}">Report this!</a>`,
+        filePath,
+        range: rangeFromLineNumber(textEditor, 0),
+        trace: [
+          {
+            type: 'Trace',
+            text: `Original message: ${ruleId} - ${message}`,
+            filePath,
+            severity: 'info',
+          },
+          {
+            type: 'Trace',
+            text: rangeText,
+            filePath,
+            severity: 'info',
+          },
+        ]
+      }
+    }
+
+    return ret
+  }))
 }
